Lazy evaluation of skills and traits
Switching to lazy evaluation of skills - no need to store the skills in
JSON, just write them in. Also means we don't need a complete database
of everything ever from the outset.
`
Added a way to get a specific lifepath by its name and setting.

Decoupled the Lifepath and Skill/Trait classes.

Made the aggregate lists private so they can't be interfered with (unit
testing should verify this).